fix extra col handle after adding first row
